price, etc to integer

+++ migrations/20250528104818-create-product.js
@@ -16,8 +16,7 @@ module.exports = {
         type: Sequelize.TEXT,
       },
       price: {
-        type: Sequelize.DECIMAL,
-        // type: Sequelize.INTEGER,
+        type: Sequelize.INTEGER,
       },
       imageUrl: {
         type: Sequelize.STRING,
